+++ my-app/src/components/championModal/championModal.scss
