+++ my-app/src/pages/watchlistPage/watchlistPage.scss
